- A lot of shader cleanups. Ifdef everything that we don't use anymore so that we produce clean shader code :D - Added a workaround for AMD bug with textureOffset + sampler2DArrayShadow not working - Changed draw order so that terrain is drawn close to the end (shading the terrain is really more expensive than the other geometry)

+++ assets/shaders/GLSL/fragmentAtoms/shadowMapping.frag
@@ -10,12 +10,14 @@
 #endif
 
 #if !defined(DISABLE_SHADOW_MAPPING)
+
 layout(binding = SHADOW_SINGLE_MAP_ARRAY)  uniform sampler2DArrayShadow    texDepthMapFromLight;
 layout(binding = SHADOW_CUBE_MAP_ARRAY)    uniform samplerCubeArrayShadow  texDepthMapFromLightCube;
 layout(binding = SHADOW_LAYERED_MAP_ARRAY) uniform sampler2DArray          texDepthMapFromLightArray;
 
 //layout(binding = TEXTURE_PREPASS_SHADOWS)  uniform sampler2D               texDepthMapFromPrePass;
 
+#include "shadowUtils.frag"
 #include "shadow_directional.frag"
 #include "shadow_point.frag"
 #include "shadow_spot.frag"
@@ -31,8 +33,8 @@ float getShadowFactorInternal(int idx) {
     if (idx >= 0 && idx < MAX_SHADOW_CASTING_LIGHTS) {
         uint lightIndex = uint(idx);
 
-        const uvec4 crtDetails = dvd_shadowLightDetails[idx];
-        switch (crtDetails.x) {
+        const vec4 crtDetails = dvd_shadowLightDetails[idx];
+        switch (uint(crtDetails.x)) {
             case LIGHT_DIRECTIONAL: return applyShadowDirectional(lightIndex, crtDetails);
             case LIGHT_OMNIDIRECTIONAL: return applyShadowPoint(lightIndex, crtDetails);
             case LIGHT_SPOT: return applyShadowSpot(lightIndex, crtDetails);
@@ -54,8 +56,8 @@ int getShadowData() {
 
 #if !defined(DISABLE_SHADOW_MAPPING)
     for (uint i = 0; i < MAX_SHADOW_CASTING_LIGHTS; ++i) {
-        const uvec4 crtDetails = dvd_shadowLightDetails[i];
-        switch (crtDetails.x) {
+        const vec4 crtDetails = dvd_shadowLightDetails[i];
+        switch (uint(crtDetails.x)) {
             case LIGHT_DIRECTIONAL: return getCSMSlice(i);
         }
     }
